+++ starknet_py/tests/e2e/client/gateway_test.py
@@ -1,6 +1,11 @@
 import pytest
+from starkware.starknet.public.abi import get_storage_var_address
 
-from starknet_py.net.client_models import TransactionStatusResponse, TransactionStatus
+from starknet_py.net.client_models import (
+    TransactionStatusResponse,
+    TransactionStatus,
+    DeployedContract,
+)
 from starknet_py.net.gateway_client import GatewayClient
 from starknet_py.net.networks import TESTNET, MAINNET, CustomGatewayUrls
 
@@ -94,3 +99,33 @@ def test_creating_client_with_custom_net_dict():
     assert gateway_client.net == net
     assert gateway_client._feeder_gateway_client.url == net["feeder_gateway_url"]
     assert gateway_client._gateway_client.url == net["gateway_url"]
+
+
+@pytest.mark.asyncio
+async def test_state_update_gateway_client(
+    gateway_client,
+    block_with_deploy_number,
+    contract_address,
+    class_hash,
+):
+    state_update = await gateway_client.get_state_update(
+        block_number=block_with_deploy_number
+    )
+
+    assert (
+        DeployedContract(
+            address=contract_address,
+            class_hash=class_hash,
+        )
+        in state_update.deployed_contracts
+    )
+
+
+@pytest.mark.asyncio
+async def test_get_storage_at_incorrect_address_gateway_client(gateway_client):
+    storage = await gateway_client.get_storage_at(
+        contract_address=0x1111,
+        key=get_storage_var_address("balance"),
+        block_hash="latest",
+    )
+    assert storage == 0
