+++ starknet_py/tests/e2e/docs/quickstart/test_using_full_node_client.py
@@ -8,6 +8,8 @@ from starknet_py.net.networks import TESTNET
 @pytest.mark.asyncio
 async def test_using_full_node_client(full_node_client, map_contract):
     # pylint: disable=import-outside-toplevel, unused-variable
+    full_node_client_fixture = full_node_client
+
     # add to docs: start
     from starknet_py.net.full_node_client import FullNodeClient
 
@@ -19,6 +21,7 @@ async def test_using_full_node_client(full_node_client, map_contract):
         max_fee=int(1e20)
     )
 
+    full_node_client = full_node_client_fixture
     # add to docs: start
 
     call_result = await full_node_client.get_block(block_number=1)
